fix imports for hover info
