+++ packages/language-support/src/index.ts
@@ -39,6 +39,7 @@ export type {
   Neo4jFunction,
   Neo4jProcedure,
   SymbolTable,
+  HoverInfo,
 } from './types.js';
 export { CypherLexer, CypherParser, CypherParserListener, CypherParserVisitor };
